KGC all double3, fixed KGC application

+++ Source/JSphGpu.h
@@ -221,7 +221,7 @@ protected:
   unsigned *MirrorCellg;
 	double3 *MirrorPosg;
   double3 *dWxCorrg; //Kernel correction in the x direction
-  float3 *dWyCorrg; //Kernel correction in the y direction
+  double3 *dWyCorrg; //Kernel correction in the y direction
   double3 *dWzCorrg; //Kernel correction in the z direction
   double4 *MLSg;
 	double3 *Velocity;
